new structure, spray-wrtc works
