+++ example/foglet.html
@@ -85,7 +85,7 @@
     <!-- Include all compiled plugins (below), or include individual files as needed -->
     <script src="bootstrap/js/bootstrap.min.js"></script>
 
-    <script src="./../dist/foglet.bundle.min.js"></script>
+    <script src="./../dist/foglet.bundle.js"></script>
     <script src="./p2p-graph.min.js"></script>
     <script src="./foglet.js"></script>
   </body>
