--- src/adapter/README.md
@@ -1,13 +1,0 @@
-# Network Adpaters
-
-Here you can find any kind of networks built on top of n2n-overlay-wrtc and neighborhood-wrtc
-
-If you want to create your own network. Refer to these modules:
-* n2n-overlay-wrtc https://github.com/RAN3D/n2n-overlay-wrtc
-* neighborhood-wrtc https://github.com/RAN3D/neighborhood-wrtc
-
-All networks built here respect the Abstract Adapter (AbstractAdapter.js) API.
-
-Hope you will do the same !   
-
-![Daftpunktocat-Thomas](https://octodex.github.com/images/daftpunktocat-thomas.gif)
